+++ src/client/client.cpp
@@ -1142,7 +1142,7 @@ void Client::sendInit(const std::string &playerName)
 	NetworkPacket pkt(TOSERVER_INIT, 1 + 2 + 2 + (1 + playerName.size()));
 
 	pkt << (u8) SER_FMT_VER_HIGHEST_READ << (u16) 0;
-	pkt << (u16) CLIENT_PROTOCOL_VERSION_MIN << (u16) CLIENT_PROTOCOL_VERSION_MAX;
+	pkt << CLIENT_PROTOCOL_VERSION_MIN << LATEST_PROTOCOL_VERSION;
 	pkt << playerName;
 
 	Send(&pkt);
