+++ src/main.cpp
@@ -728,7 +728,7 @@ static void startup_message()
 	print_version(infostream);
 	infostream << "SER_FMT_VER_HIGHEST_READ=" <<
 		TOSTRING(SER_FMT_VER_HIGHEST_READ) <<
-		" LATEST_PROTOCOL_VERSION=" << TOSTRING(LATEST_PROTOCOL_VERSION)
+		" LATEST_PROTOCOL_VERSION=" << LATEST_PROTOCOL_VERSION
 		<< std::endl;
 }
