+++ src/network/CMakeLists.txt
@@ -4,6 +4,7 @@ set(common_network_SRCS
 	${CMAKE_CURRENT_SOURCE_DIR}/mtp/impl.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/mtp/threads.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/networkpacket.cpp
+	${CMAKE_CURRENT_SOURCE_DIR}/networkprotocol.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/serveropcodes.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/serverpackethandler.cpp
 	${CMAKE_CURRENT_SOURCE_DIR}/socket.cpp
